Refactor code
Add core folder with logics inside it
Improve naming of utils.py
