+++ backend/api/endpoints.py
@@ -3,7 +3,7 @@ import os
 
 from fastapi import APIRouter
 
-import utils.openai_wrapper as openai_wrapper
+import core.ai_service as ai_service
 from models.job import Job
 from models.latex import Latex
 from models.profile import Profile, Resume
@@ -15,7 +15,7 @@ from models.templates import (
     john_doe_preferences,
     john_doe_resume,
 )
-from utils.utils import generate_pdf_from_latex, save_pdf
+from utils.file_ops import generate_pdf_from_latex, save_pdf
 
 router = APIRouter(prefix="/api")
 
@@ -24,7 +24,7 @@ def determine_eligibility(job: Job, profile: Profile = Profile(resume=Resume(joh
     """
     Gets profile and job description and determines eligibility for applying to the job
     """
-    eligibility, reason = openai_wrapper.consider_eligibility(job.description, profile.legal_authorization, tailoring_options.ai_model)
+    eligibility, reason = ai_service.consider_eligibility(job.description, profile.legal_authorization, tailoring_options.ai_model)
     return {
         "eligibility": eligibility,
         "reason": reason
@@ -35,7 +35,7 @@ def determine_suitability(job: Job, profile: Profile = Profile(resume=Resume(joh
     """
     Gets profile and job description and determines suitability for applying to the job 
     """
-    suitability, reason = openai_wrapper.consider_suitability(job.description, profile.preferences, tailoring_options.ai_model)
+    suitability, reason = ai_service.consider_suitability(job.description, profile.preferences, tailoring_options.ai_model)
     return {
         "suitability": suitability,
         "reason": reason
@@ -46,14 +46,14 @@ def generate_tailored_plain_resume(job: Job, profile: Profile = Profile(resume=R
     """
     Gets resume and job description in plain text and returns tailored resume as a string
     """
-    return openai_wrapper.create_tailored_plain_resume(profile.resume.text, job.description, tailoring_options.ai_model, tailoring_options.resume_template)
+    return ai_service.create_tailored_plain_resume(profile.resume.text, job.description, tailoring_options.ai_model, tailoring_options.resume_template)
 
 @router.post("/generate-tailored-plain-coverletter")
 def generate_tailored_plain_coverletter(job: Job, profile: Profile = Profile(Resume(john_doe_resume)), tailoring_options: TailoringOptions = TailoringOptions()) -> str:  # We should not pass tailoring options everytime, should be a config for each user. It could be kept with a session for example.
     """
     Gets resume and job description in plain text and returns customized cover letter as a string
     """
-    return openai_wrapper.create_tailored_plain_coverletter(profile.resume.text, job.description, tailoring_options.ai_model)
+    return ai_service.create_tailored_plain_coverletter(profile.resume.text, job.description, tailoring_options.ai_model)
 
 @router.post("/generate-latex-resume-save")
 def generate_tailored_latex_resume_save(job: Job, profile: Profile = Profile(Resume(john_doe_resume)), tailoring_options: TailoringOptions = TailoringOptions()):
@@ -61,14 +61,14 @@ def generate_tailored_latex_resume_save(job: Job, profile: Profile = Profile(Res
     Gets resume and job description in plain text and saves tailored resume
     """
     tailored_plain_resume = generate_tailored_plain_resume(job, profile, tailoring_options)
-    company_name = openai_wrapper.ai_prompt(
+    company_name = ai_service.ai_prompt(
         f"Give the name of the company that this job description is for. As the output just give the name, nothing else. Job description: {job.description}"
     )  # Since this is a simple task we use the cheapest ai
     
     current_time = datetime.now().strftime("%Y-%m-%d_%H-%M-%S")
     pdf_path = f'./Resumes/{current_time}_{company_name}'
     os.makedirs(pdf_path, exist_ok=True)
-    latex_compiler_response, latex_code = openai_wrapper.covert_plain_resume_to_latex(
+    latex_compiler_response, latex_code = ai_service.covert_plain_resume_to_latex(
         current_time,
         company_name,
         tailored_plain_resume,
@@ -93,7 +93,7 @@ def answer_application_questions(
     """
     Gets resume, job description, and questions in the job applicaton and answer to them based on resume and job description.
     """
-    return openai_wrapper.generate_answer_questions(
+    return ai_service.generate_answer_questions(
         profile.resume.text,
         job.description,
         question.description,
